geo types tests work
