+++ packages/xlucene-evaluator/src/document-matcher/index.ts
@@ -1,33 +1,10 @@
+
 import _ from 'lodash';
 import { debugLogger } from '@terascope/utils';
-// @ts-ignore
-import { both, either, not, identity, allPass, anyPass } from 'rambda';
-import {
-    Parser,
-    AST,
-    isTerm,
-    isExists,
-    // @ts-ignore
-
-    isRange,
-    // @ts-ignore
-    isConjunction,
-    isLogicalGroup,
-    getAnyValue,
-    getField,
-    isNegation,
-    isFieldGroup,
-    Range
-} from '../parser';
-
-// import TypeManager from './type-manager';
-
-import {
-    bindThis,
-    checkValue
-} from '../utils';
-
-import { TypeConfig, BooleanCB } from '../interfaces';
+import { Parser } from '../parser';
+import { bindThis } from '../utils';
+import { TypeConfig } from '../interfaces';
+import logicBuilder from './logic-builder';
 
 // @ts-ignore
 const logger = debugLogger('document-matcher');
@@ -44,11 +21,11 @@ export default class DocumentMatcher {
             this.parse(luceneStr);
         }
     }
-
+    // TODO: remove this method
     public parse(luceneStr: string, typeConfig?: TypeConfig):void {
         const parser = new Parser(luceneStr);
         const types = typeConfig || this.typeConfig;
-        const resultingFN = buildLogicFn(parser, types);
+        const resultingFN = logicBuilder(parser, types);
         this.filterFn = resultingFN;
     }
 
@@ -57,144 +34,3 @@ export default class DocumentMatcher {
         return this.filterFn(doc);
     }
 }
-
-const negate = (fn:any) => (data:any) => not(fn(data));
-
-const logicNode = (boolFn: BooleanCB, node:AST) => {
-    if (isNegation(node)) return negate(boolFn);
-    return boolFn;
-};
-
-function buildLogicFn(parser: Parser, typeConfig: TypeConfig|undefined) {
-    // const types = new TypeManager(parser, typeConfig);
-    // const parsedAst = types.processAst();
-    console.log('original ast', JSON.stringify(parser.ast, null, 4));
-    function walkAst(node: AST): BooleanCB {
-        let fnResults;
-        const value = getAnyValue(node);
-        const field = getField(node);
-
-        if (isNegation(node)) {
-            // @ts-ignore
-            const childLogic = walkAst(node.node);
-            fnResults = negate(childLogic);
-        }
-
-        // TODO: Deal with regex and wildcard
-        if (isTerm(node)) {
-            const isValue = (data: any) => {
-                return data === value;
-            };
-            const fn = checkValue(field as string, isValue);
-            fnResults = logicNode(fn, node);
-        }
-
-        if (isExists(node)) {
-            const valueExists = (value: any) => value != null;
-            const fn = checkValue(field, valueExists);
-            fnResults = logicNode(fn, node);
-        }
-
-        if (isRange(node)) {
-            const fn = checkValue(field, rangeFn(node));
-            fnResults = logicNode(fn, node);
-        }
-
-        // if (isRange(node)) {
-        //     const fn = checkValue(node.field, parseRange(node));
-        //     fnResults = logicNode(fn, node);
-        // }
-
-        // if (isParsedNode(node)) {
-        //     const fn  = node.callback;
-        //     fnResults = logicNode(fn, node.negated);
-        // }
-
-        if (isLogicalGroup(node) || isFieldGroup(node)) {
-            const logicGroups: BooleanCB[] = [];
-
-            node.flow.forEach(conjunction => {
-                const conjunctionRules = conjunction.nodes.map(node => walkAst(node));
-                logicGroups.push(allPass(conjunctionRules));
-            });
-
-            fnResults = logicNode(anyPass(logicGroups), node);
-        }
-
-        if (!fnResults) fnResults = () => false;
-        return fnResults;
-    }
-
-    return walkAst(parser.ast);
-}
-
-function rangeFn(node: Range): BooleanCB {
-    const mapping = {
-        gte: _.gte,
-        gt: _.gt,
-        lte: _.lte,
-        lt: _.lt
-    };
-    const { left, right } = node;
-
-    if (!right) {
-        return (data: any) => mapping[left.operator](data, left.value);
-    }
-
-    return (data: any) => mapping[left.operator](data, left.value) && mapping[right.operator](data, right.value);
-}
-
-// function parseRange(node: RangeAST) {
-//     const {
-//         inclusive_min: incMin,
-//         inclusive_max: incMax,
-//     } = node;
-//     let { term_min: minValue, term_max: maxValue } = node;
-//     let rangeFn = (_data: any) => false;
-
-//     if (minValue === '*') minValue = Number.NEGATIVE_INFINITY;
-//     if (maxValue === '*') maxValue = Number.POSITIVE_INFINITY;
-
-//     // IPs can use range syntax, if no type is set it needs to return
-//     // a hard coded string interpolated value
-//     [minValue, maxValue] = [minValue, maxValue].map((data) => {
-//         if (typeof data === 'string') return `"${data}"`;
-//         return data;
-//     });
-
-//     // ie age:>10 || age:(>10 AND <=20)
-//     if (!incMin && incMax) {
-//         if (isInfiniteMax(maxValue)) {
-//             rangeFn = (data: any) => _.gt(data, minValue);
-//         } else {
-//             rangeFn = (data: any) => _.gt(data, minValue) && _.lte(data, maxValue);
-//         }
-//     }
-
-//     // ie age:<10 || age:(<=10 AND >20)
-//     if (incMin && !incMax) {
-//         if (isInfiniteMin(minValue)) {
-//             rangeFn = (data: any) => _.lt(data, maxValue);
-//         } else {
-//             rangeFn = (data: any) => _.lt(data, maxValue) && _.gte(data, minValue);
-//         }
-//     }
-
-//     // ie age:<=10, age:>=10, age:(>=10 AND <=20)
-//     if (incMin && incMax) {
-//         if (isInfiniteMax(maxValue)) {
-//             rangeFn = (data: any) => _.gte(data, minValue);
-//         } else if (isInfiniteMin(minValue)) {
-//             rangeFn = (data: any) => _.lte(data, maxValue);
-//         } else {
-//             rangeFn = (data: any) => _.lte(data, maxValue) && _.gte(data, minValue);
-//         }
-//     }
-
-//     // ie age:(>10 AND <20)
-//     if (!incMin && !incMax) {
-//         rangeFn = (data: any) => _.lt(data, maxValue) && _.gt(data, minValue);
-//     }
-
-//     return rangeFn;
-// }
